refactored references to player, esp update function
